dj-api/pyproject.toml
[tool.poetry]
name = "dj-api"
version = "0.1.0"
description = ""
authors = ["Wei Hung <wayne930242@gmail.com>"]
readme = "README.md"
packages = [{include = "dj_api"}]

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.5.0"
django = "4.0.3"
django-cors-headers = "3.11.0"
djangorestframework = "3.13.1"
pytz = "2021.3"
sqlparse = "0.4.2"
tzdata = "2021.5"
langchain = "^0.0.149"
openai = "^0.27.4"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
